Kai dev

+++ Coursework2.X/Delay.h
@@ -12,7 +12,7 @@
 extern "C" {
 #endif
 
-    void Delay(int n);
+    void Delay(unsigned short n);
     void therm_delay(char, char);
